+++ src/appleseed/CMakeLists.txt
@@ -263,7 +263,6 @@ set (foundation_math_sources
     foundation/math/knn.h
     foundation/math/matrix.h
     foundation/math/microfacet.h
-    foundation/math/microfacet2.h
     foundation/math/minmax.h
     foundation/math/mis.h
     foundation/math/noise.cpp
@@ -346,7 +345,7 @@ set (foundation_meta_benchmarks_sources
     foundation/meta/benchmarks/benchmark_knn.cpp
     foundation/meta/benchmarks/benchmark_math_filter.cpp
     foundation/meta/benchmarks/benchmark_matrix.cpp
-    foundation/meta/benchmarks/benchmark_microfacet.cpp
+    #foundation/meta/benchmarks/benchmark_microfacet.cpp
     foundation/meta/benchmarks/benchmark_permutation.cpp
     foundation/meta/benchmarks/benchmark_poolallocator.cpp
     foundation/meta/benchmarks/benchmark_qmc.cpp
@@ -410,7 +409,6 @@ set (foundation_meta_tests_sources
     foundation/meta/tests/test_matrix.cpp
     foundation/meta/tests/test_memory.cpp
     foundation/meta/tests/test_microfacet.cpp
-    foundation/meta/tests/test_microfacet2.cpp
     foundation/meta/tests/test_minmax.cpp
     foundation/meta/tests/test_noise.cpp
     foundation/meta/tests/test_objmeshfilereader.cpp
